Refactoring website next into website-honestly WIP
