+++ package.json
@@ -41,6 +41,7 @@
     "file-loader": "^0.9.0",
     "flow-bin": "^0.36.0",
     "flow-typed": "^2.0.0",
+    "html-to-react": "^1.2.4",
     "html-webpack-plugin": "^2.22.0",
     "http-server": "^0.9.0",
     "imagemin-cli": "^3.0.0",
@@ -66,7 +67,9 @@
     "react-addons-test-utils": "^15.3.2",
     "react-dom": "^15.3.1",
     "react-motion": "^0.4.7",
+    "react-redux": "^5.0.3",
     "react-swipeable-views": "^0.10.1",
+    "redux": "^3.6.0",
     "reset-css": "^2.2.0",
     "serverless": "1.0.0-rc.1",
     "style-loader": "^0.13.1",
