+++ site/components/fonts/style.css
@@ -1,0 +1,31 @@
+/*
+* This component is ported from Website Next and is meant to be
+* used only by Jobs and Events components
+*/
+
+@value font from "../variables/colors.css";
+
+/*
+ * Note: These currently do not work 100% with post-css module values. It looks
+ * like it is in the pipeline but for now these values will be parsed to their
+ * first declared value. i.e. "CenturyGothic"
+ * (https://github.com/css-modules/postcss-modules-values/pull/23)
+ */
+
+@value default: "ProximaNova-Light", Helvetica, Arial, sans-serif;
+@value hero: "Sentinel A", "Sentinel B", Georgia, serif;
+
+.defaultFontSize {
+  font-size: 16px;
+  line-height: 1.4;
+}
+
+.defaultFont {
+  font-family: default;
+  color: font;
+}
+
+.heroFont {
+  font-family: hero;
+  color: font;
+}
