+++ site/components/html-parser/style.css
@@ -1,7 +1,7 @@
 @value font from "../variables/colors.css";
 
 .rawHtml {
-  composes: sansSerif from "../../../../../site/css/typography/_fonts.css";
+  composes: sansSerif from "../../css/typography/_fonts.css";
 }
 
 .rawHtml strong {
