+++ site/components/note/index.js
